fix routes and refactor
