+++ ui/src/features/WorkspaceLeftPanel/components/BottomSection/components/index.ts
@@ -1,3 +1,1 @@
 export * from "./WorkspaceSettings";
-export * from "./DeploymentManager";
-export * from "./ProjectManager";
