+++ ui/src/lib/gql/__gen__/gql.ts
@@ -14,7 +14,7 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
  * Learn more about it here: https://the-guild.dev/graphql/codegen/plugins/presets/preset-client#reducing-bundle-size
  */
 type Documents = {
-    "mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}": typeof types.CreateDeploymentDocument,
+    "mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetDeploymentHead($input: GetHeadInput!) {\n  deploymentHead(input: $input) {\n    ...Deployment\n  }\n}": typeof types.CreateDeploymentDocument,
     "fragment Project on Project {\n  id\n  name\n  description\n  createdAt\n  updatedAt\n  workspaceId\n  sharedToken\n  deployment {\n    ...Deployment\n  }\n}\n\nfragment Deployment on Deployment {\n  id\n  projectId\n  workspaceId\n  workflowUrl\n  description\n  version\n  createdAt\n  updatedAt\n  project {\n    name\n  }\n}\n\nfragment Trigger on Trigger {\n  id\n  createdAt\n  updatedAt\n  lastTriggered\n  workspaceId\n  deploymentId\n  deployment {\n    ...Deployment\n  }\n  eventSource\n  authToken\n  timeInterval\n  description\n}\n\nfragment Job on Job {\n  id\n  workspaceId\n  status\n  startedAt\n  completedAt\n  logsURL\n  outputURLs\n  deployment {\n    id\n    description\n  }\n}\n\nfragment Log on Log {\n  jobId\n  nodeId\n  timestamp\n  logLevel\n  message\n}": typeof types.ProjectFragmentDoc,
     "query GetJobs($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  jobs(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Job\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetJob($id: ID!) {\n  job(id: $id) {\n    ...Job\n  }\n}\n\nmutation CancelJob($input: CancelJobInput!) {\n  cancelJob(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}": typeof types.GetJobsDocument,
     "mutation CreateProject($input: CreateProjectInput!) {\n  createProject(input: $input) {\n    project {\n      ...Project\n    }\n  }\n}\n\nquery GetProjects($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  projects(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Project\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetProjectById($projectId: ID!) {\n  node(id: $projectId, type: PROJECT) {\n    __typename\n    ...Project\n  }\n}\n\nmutation UpdateProject($input: UpdateProjectInput!) {\n  updateProject(input: $input) {\n    project {\n      ...Project\n    }\n  }\n}\n\nmutation DeleteProject($input: DeleteProjectInput!) {\n  deleteProject(input: $input) {\n    projectId\n  }\n}\n\nmutation RunProject($input: RunProjectInput!) {\n  runProject(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}": typeof types.CreateProjectDocument,
@@ -25,7 +25,7 @@ type Documents = {
     "fragment Workspace on Workspace {\n  id\n  name\n  personal\n  members {\n    userId\n    role\n    user {\n      id\n      email\n      name\n    }\n  }\n}\n\nmutation CreateWorkspace($input: CreateWorkspaceInput!) {\n  createWorkspace(input: $input) {\n    workspace {\n      ...Workspace\n    }\n  }\n}\n\nquery GetWorkspaces {\n  me {\n    id\n    workspaces {\n      ...Workspace\n    }\n  }\n}\n\nquery GetWorkspaceById($workspaceId: ID!) {\n  node(id: $workspaceId, type: WORKSPACE) {\n    __typename\n    ...Workspace\n  }\n}\n\nmutation UpdateWorkspace($input: UpdateWorkspaceInput!) {\n  updateWorkspace(input: $input) {\n    workspace {\n      ...Workspace\n    }\n  }\n}\n\nmutation DeleteWorkspace($input: DeleteWorkspaceInput!) {\n  deleteWorkspace(input: $input) {\n    workspaceId\n  }\n}\n\nmutation AddMemberToWorkspace($input: AddMemberToWorkspaceInput!) {\n  addMemberToWorkspace(input: $input) {\n    workspace {\n      ...Workspace\n    }\n  }\n}\n\nmutation RemoveMemberFromWorkspace($input: RemoveMemberFromWorkspaceInput!) {\n  removeMemberFromWorkspace(input: $input) {\n    workspace {\n      ...Workspace\n    }\n  }\n}\n\nmutation UpdateMemberOfWorkspace($input: UpdateMemberOfWorkspaceInput!) {\n  updateMemberOfWorkspace(input: $input) {\n    workspace {\n      ...Workspace\n    }\n  }\n}": typeof types.WorkspaceFragmentDoc,
 };
 const documents: Documents = {
-    "mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}": types.CreateDeploymentDocument,
+    "mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetDeploymentHead($input: GetHeadInput!) {\n  deploymentHead(input: $input) {\n    ...Deployment\n  }\n}": types.CreateDeploymentDocument,
     "fragment Project on Project {\n  id\n  name\n  description\n  createdAt\n  updatedAt\n  workspaceId\n  sharedToken\n  deployment {\n    ...Deployment\n  }\n}\n\nfragment Deployment on Deployment {\n  id\n  projectId\n  workspaceId\n  workflowUrl\n  description\n  version\n  createdAt\n  updatedAt\n  project {\n    name\n  }\n}\n\nfragment Trigger on Trigger {\n  id\n  createdAt\n  updatedAt\n  lastTriggered\n  workspaceId\n  deploymentId\n  deployment {\n    ...Deployment\n  }\n  eventSource\n  authToken\n  timeInterval\n  description\n}\n\nfragment Job on Job {\n  id\n  workspaceId\n  status\n  startedAt\n  completedAt\n  logsURL\n  outputURLs\n  deployment {\n    id\n    description\n  }\n}\n\nfragment Log on Log {\n  jobId\n  nodeId\n  timestamp\n  logLevel\n  message\n}": types.ProjectFragmentDoc,
     "query GetJobs($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  jobs(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Job\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetJob($id: ID!) {\n  job(id: $id) {\n    ...Job\n  }\n}\n\nmutation CancelJob($input: CancelJobInput!) {\n  cancelJob(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}": types.GetJobsDocument,
     "mutation CreateProject($input: CreateProjectInput!) {\n  createProject(input: $input) {\n    project {\n      ...Project\n    }\n  }\n}\n\nquery GetProjects($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  projects(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Project\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetProjectById($projectId: ID!) {\n  node(id: $projectId, type: PROJECT) {\n    __typename\n    ...Project\n  }\n}\n\nmutation UpdateProject($input: UpdateProjectInput!) {\n  updateProject(input: $input) {\n    project {\n      ...Project\n    }\n  }\n}\n\nmutation DeleteProject($input: DeleteProjectInput!) {\n  deleteProject(input: $input) {\n    projectId\n  }\n}\n\nmutation RunProject($input: RunProjectInput!) {\n  runProject(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}": types.CreateProjectDocument,
@@ -53,7 +53,7 @@ export function graphql(source: string): unknown;
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}"): (typeof documents)["mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}"];
+export function graphql(source: "mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetDeploymentHead($input: GetHeadInput!) {\n  deploymentHead(input: $input) {\n    ...Deployment\n  }\n}"): (typeof documents)["mutation CreateDeployment($input: CreateDeploymentInput!) {\n  createDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation UpdateDeployment($input: UpdateDeploymentInput!) {\n  updateDeployment(input: $input) {\n    deployment {\n      ...Deployment\n    }\n  }\n}\n\nmutation DeleteDeployment($input: DeleteDeploymentInput!) {\n  deleteDeployment(input: $input) {\n    deploymentId\n  }\n}\n\nmutation ExecuteDeployment($input: ExecuteDeploymentInput!) {\n  executeDeployment(input: $input) {\n    job {\n      ...Job\n    }\n  }\n}\n\nquery GetDeployments($workspaceId: ID!, $pagination: PageBasedPagination!) {\n  deployments(workspaceId: $workspaceId, pagination: $pagination) {\n    totalCount\n    nodes {\n      ...Deployment\n    }\n    pageInfo {\n      totalCount\n      currentPage\n      totalPages\n    }\n  }\n}\n\nquery GetDeploymentHead($input: GetHeadInput!) {\n  deploymentHead(input: $input) {\n    ...Deployment\n  }\n}"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
