+++ ui/src/routes/workspaces.$workspaceId.deployments.tsx
@@ -2,8 +2,6 @@ import { createFileRoute } from "@tanstack/react-router";
 
 import { DeploymentManager } from "@flow/features/WorkspaceDeployments";
 
-export const Route = createFileRoute(
-  "/workspaces/$workspaceId/deployments/$tab",
-)({
+export const Route = createFileRoute("/workspaces/$workspaceId/deployments")({
   component: () => <DeploymentManager />,
 });
